+++ stories/VerticalNavigation.js
@@ -21,7 +21,6 @@ const searchInput = (
     label="search"
     placeholder="Quick Search"
     type="text"
-    value={undefined}
   />
 );
 
@@ -35,33 +34,31 @@ stories
       <VerticalNavigationSection title={text('Title section 1', 'Reports')} id="entity-header">
         <VerticalNavigationItem
           href="#"
-          isActive={boolean('Section 1 Item 1 isActive', true) || undefined}
-          notification={number('Notification section 1 item 1', null) || undefined}
-          notificationLabel={text('NotificationLabel section 1 item 1', '') || undefined}
+          isActive={boolean('Section 1 Item 1 isActive', true)}
+          notification={text('Notification section 1 item 1', 'None')}
+          notificationLabel={text('NotificationLabel section 1 item 1', '')}
         >Recent</VerticalNavigationItem>
         <VerticalNavigationItem
           href="#"
-          isActive={boolean('Section 1 Item 2 isActive', false) || undefined}
-          notification={number('Notification section 1 item 2', 3) || undefined}
-          notificationLabel={text('NotificationLabel section 1 item 2', 'new items') || undefined}
+          isActive={boolean('Section 1 Item 2 isActive', false)}
+          notification={number('Notification section 1 item 2', 3)}
+          notificationLabel={text('NotificationLabel section 1 item 2', 'new items')}
         >Created By Me</VerticalNavigationItem>
       </VerticalNavigationSection>
       <VerticalNavigationSection title={text('Title section 2', 'Folders')} id="folder-header">
         <VerticalNavigationItem
           icon={object('Icon 1', fileIcon)}
           onClick={action('clicked file 1')}
-          isActive={boolean('Section 2 Item 1 isActive', false) || undefined}
-          notification={number('Notification section 2 item 1', null) || undefined}
-          notificationLabel={text('NotificationLabel section 2 item 1', '') || undefined}
+          isActive={boolean('Section 2 Item 1 isActive', false)}
         >
           Created By Me
         </VerticalNavigationItem>
         <VerticalNavigationItem
           icon={object('Icon 2', fileIcon)}
           onClick={action('clicked file 2')}
-          isActive={boolean('Section 2 Item 2 isActive', false) || undefined}
-          notification={number('Notification section 2 item 2', null) || undefined}
-          notificationLabel={text('NotificationLabel section 2 item 2', '') || undefined}
+          isActive={boolean('Section 2 Item 2 isActive', false)}
+          notification={number('Notification section 1 item 2', 3)}
+          notificationLabel={text('NotificationLabel section 1 item 2', 'new items')}
         >
           Shared with me
         </VerticalNavigationItem>